requirements.txt
django
gunicorn
django-heroku
APScheduler==3.0.0
rq==1.5.0
facebook-sdk==3.1.0
opencv-python==4.2.0.34
numpy==1.14.0

